+++ server/src/main/java/io/seata/server/session/SessionHelper.java
@@ -167,7 +167,7 @@ public class SessionHelper {
             GlobalStatus currentStatus = globalSession.getStatus();
             boolean retryBranch =
                 currentStatus == GlobalStatus.TimeoutRollbackRetrying || currentStatus == GlobalStatus.RollbackRetrying;
-            if (isTimeoutGlobalStatus(currentStatus)) {
+            if (SessionStatusHelper.isTimeoutGlobalStatus(currentStatus)) {
                 globalSession.changeGlobalStatus(GlobalStatus.TimeoutRollbacked);
             } else {
                 globalSession.changeGlobalStatus(GlobalStatus.Rollbacked);
@@ -192,7 +192,7 @@ public class SessionHelper {
      */
     public static void endRollbackFailed(GlobalSession globalSession, boolean retryGlobal) throws TransactionException {
         GlobalStatus currentStatus = globalSession.getStatus();
-        if (isTimeoutGlobalStatus(currentStatus)) {
+        if (SessionStatusHelper.isTimeoutGlobalStatus(currentStatus)) {
             globalSession.changeGlobalStatus(GlobalStatus.TimeoutRollbackFailed);
         } else {
             globalSession.changeGlobalStatus(GlobalStatus.RollbackFailed);
@@ -202,66 +202,7 @@ public class SessionHelper {
         MetricsPublisher.postSessionDoneEvent(globalSession, retryGlobal, false);
     }
 
-    /**
-     * is timeout global status timeout
-     *
-     * @param status the global session
-     */
-    public static boolean isTimeoutGlobalStatus(GlobalStatus status) {
-        return status == GlobalStatus.TimeoutRollbacked
-                || status == GlobalStatus.TimeoutRollbackFailed
-                || status == GlobalStatus.TimeoutRollbacking
-                || status == GlobalStatus.TimeoutRollbackRetrying;
-    }
 
-    /**
-     * is rollback global status timeout
-     *
-     * @param status the global session
-     */
-    public static boolean isRollbackGlobalStatus(GlobalStatus status) {
-        return status == GlobalStatus.Rollbacking
-                || status == GlobalStatus.RollbackRetrying
-                || status == GlobalStatus.Rollbacked
-                || status == GlobalStatus.RollbackFailed
-                || status == GlobalStatus.RollbackRetryTimeout;
-    }
-
-    /**
-     * is commit global status timeout
-     *
-     * @param status the global session
-     */
-    public static boolean isCommitGlobalStatus(GlobalStatus status) {
-        return status == GlobalStatus.Committing
-                || status == GlobalStatus.AsyncCommitting
-                || status == GlobalStatus.CommitRetrying
-                || status == GlobalStatus.Committed
-                || status == GlobalStatus.CommitFailed
-                || status == GlobalStatus.CommitRetryTimeout;
-    }
-
-    /**
-     * check the relation of before status and after status
-     *
-     * @param before the global session
-     * @param after the global session
-     */
-    public static boolean validateUpdateStatus(GlobalStatus before, GlobalStatus after) {
-        if (isTimeoutGlobalStatus(before) && isCommitGlobalStatus(after)) {
-            return false;
-        }
-        if (isCommitGlobalStatus(before) && isTimeoutGlobalStatus(after)) {
-            return false;
-        }
-        if (isRollbackGlobalStatus(before) && isCommitGlobalStatus(after)) {
-            return false;
-        }
-        if (isCommitGlobalStatus(before) && isRollbackGlobalStatus(after)) {
-            return false;
-        }
-        return true;
-    }
 
     /**
      * Foreach global sessions.
