+++ server/src/main/java/io/seata/server/storage/redis/store/RedisTransactionStoreManager.java
@@ -28,7 +28,7 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 import io.seata.config.Configuration;
 import io.seata.config.ConfigurationFactory;
-import io.seata.server.session.SessionHelper;
+import io.seata.server.session.SessionStatusHelper;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import com.google.common.collect.ImmutableMap;
@@ -327,7 +327,7 @@ public class RedisTransactionStoreManager extends AbstractTransactionStoreManage
                 jedis.unwatch();
                 return true;
             }
-            if (SessionHelper.validateUpdateStatus(GlobalStatus.get(Integer.parseInt(previousStatus)),
+            if (!SessionStatusHelper.validateUpdateStatus(GlobalStatus.get(Integer.parseInt(previousStatus)),
                 GlobalStatus.get(globalTransactionDO.getStatus()))) {
                 throw new StoreException("illegal changing of global status, update global transaction failed.");
             }
